Fix range of November possibilities
The variable MMsigma/MMdrift sets a range of possibilities to integrate over. This stabilizes the jerseyvotes calculation. For long times, the range should be wider - with a floor. The max() function does that. For some reason I set it to min(). Probably not enough coffee. :-P now matches similar calculations elsewhere, for example line in in EV_prediction.m